backend/requirements.txt
flask
flask-cors
openai
python-dotenv
uvicorn
pypdf
langchain
chromadb
pytest
boto3
langchain-ollama
langchain-chroma
firebase-admin
stripe
psycopg2-binary
